feat: added a helper methord for library functions
