+++ app/src/main/java/com/piper/urbandemo/activity/SplashActivity.java
@@ -6,6 +6,7 @@ import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
 import com.piper.myappauth.activity.SigninActivity;
+import com.piper.myappauth.helper.AuthActivity;
 import com.piper.myappauth.helper.Keys;
 import com.piper.urbandemo.R;
 import com.piper.urbandemo.activity.home.HomeActivity;
@@ -23,9 +24,10 @@ public class SplashActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
 
-        SigninActivity.initialize(this);
+        //initialize the library
+        AuthActivity.initialize(this);
 
-        if (SigninActivity.isAuthenticated()) {
+        if (AuthActivity.isAuthenticated()) {
             navigateToHome();
         } else {
             //start activity for authentication
